T-RAX TEAM  HEPSİBURADA/requirements.txt
tkinter
selenium
webdriver_manager
Pillow
transformers
torch
